+++ NEWS.md
@@ -4,6 +4,7 @@
 - Add `bencapex` as illustrative data frame.
 - Add `bcai()` for calculating Benati and Capurri's (2026) alignment index. You can think of this as a kind of chance-corrected *S* score.
 - Rewrote `srs()` to function more like `bcai()`. `range` argument also replaced with `levels`. Support added for weighted, squared distances.
+- Add `download.fpsim.data()`.
 
 # fpsim 0.1.0